Fix #304: unblock decode endpoint and consolidate decoder
