+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/corescope
+
+go 1.22
